Edited README (usage section). Added paragraph in README about enabling application level encryption. Moved AES cipher from test module to be a domain service.
